+++ lib/spreedly-core-ruby.rb
@@ -12,7 +12,7 @@ require 'spreedly-core-ruby/transactions'
 require 'active_support/core_ext/hash/conversions'
 
 module SpreedlyCore
-  # Hash of user friendly credit card name to SpreedlyCore API name
+  # Hash of user friendly credit card name to Spreedly API name
   CARD_TYPES = {
     "Visa" => "visa",
     "MasterCard" => "master",
@@ -22,7 +22,7 @@ module SpreedlyCore
 
   class Error < RuntimeError; end
 
-  # Custom exception which occurs when a request to SpreedlyCore times out
+  # Custom exception which occurs when a request to Spreedly times out
   # See SpreedlyCore::Base.default_timeout
   class TimeOutError < Error; end
   class InvalidResponse < Error
@@ -42,23 +42,23 @@ module SpreedlyCore
     end
   end
 
-  # Configure SpreedlyCore with a particular account.
+  # Configure Spreedly with a particular account.
   # Strongly prefers environment variables for credentials
   # and will issue a stern warning should they not be present.
-  # Reluctantly accepts :login and :secret as options
+  # Reluctantly accepts :environment_key and :access_secret as options
   def self.configure(options = {})
-    login = ENV['SPREEDLYCORE_API_LOGIN']
-    secret = ENV['SPREEDLYCORE_API_SECRET']
+    environment_key = (ENV["SPREEDLYCORE_ENVIRONMENT_KEY"] || ENV['SPREEDLYCORE_API_LOGIN'])
+    secret = (ENV["SPREEDLYCORE_ACCESS_SECRET"] || ENV['SPREEDLYCORE_API_SECRET'])
     gateway_token = ENV['SPREEDLYCORE_GATEWAY_TOKEN']
 
-    if options[:api_login]
-      Kernel.warn("ENV and arg both present for api_login. Defaulting to arg value") if login
-      login = options[:api_login]
+    if(options[:environment_key] || options[:api_login])
+      Kernel.warn("ENV and arg both present for environment_key. Defaulting to arg value") if environment_key
+      environment_key = (options[:environment_key] || options[:api_login])
     end
 
-    if options[:api_secret]
-      Kernel.warn("ENV and arg both present for api_secret. Defaulting to arg value") if secret
-      secret = options[:api_secret]
+    if(options[:access_secret] || options[:api_secret])
+      Kernel.warn("ENV and arg both present for access_secret. Defaulting to arg value") if secret
+      secret = (options[:access_secret] || options[:api_secret])
     end
 
     if options[:gateway_token]
@@ -67,18 +67,18 @@ module SpreedlyCore
     end
     options[:gateway_token] ||= gateway_token
 
-    if options[:api_login] || options[:api_secret]
-      Kernel.warn("It is STRONGLY preferred that you house your Spreedly Core credentials only in environment variables.")
-      Kernel.warn("This gem prefers only environment variables named SPREEDLYCORE_API_LOGIN, SPREEDLYCORE_API_SECRET, and optionally SPREEDLYCORE_GATEWAY_TOKEN.")
+    if(options[:environment_key] || options[:access_secret])
+      Kernel.warn("It is STRONGLY preferred that you house your Spreedly credentials only in environment variables.")
+      Kernel.warn("This gem prefers only environment variables named SPREEDLYCORE_ENVIRONMENT_KEY, SPREEDLYCORE_ACCESS_SECRET, and optionally SPREEDLYCORE_GATEWAY_TOKEN.")
     end
 
-    if login.nil? || secret.nil?
-      raise ArgumentError.new("You must provide a login and a secret. Gem will look for ENV['SPREEDLYCORE_API_LOGIN'] and ENV['SPREEDLYCORE_API_SECRET'], but you may also pass in a hash with :api_login and :api_secret keys.")
+    if environment_key.nil? || secret.nil?
+      raise ArgumentError.new("You must provide a environment_key and a secret. Gem will look for ENV['SPREEDLYCORE_ENVIRONMENT_KEY'] and ENV['SPREEDLYCORE_ACCESS_SECRET'], but you may also pass in a hash with :environment_key and :access_secret keys.")
     end
 
-    options[:endpoint] ||= "https://spreedlycore.com/#{SpreedlyCore::API_VERSION}"
+    options[:endpoint] ||= "https://core.spreedly.com/#{SpreedlyCore::API_VERSION}"
 
-    Base.configure(login, secret, options)
+    Base.configure(environment_key, secret, options)
   end
 
   def self.gateway_token=(gateway_token)
@@ -89,13 +89,22 @@ module SpreedlyCore
     Base.gateway_token
   end
 
-  # returns the configured SpreedlyCore login
-  def self.login; Base.login; end
+  # returns the configured Spreedly environment key
+  def self.environment_key; Base.environment_key; end
 
   # A container for a response from a payment gateway
   class Response < Base
-    attr_reader(:success, :message, :avs_code, :avs_message, :cvv_code,
-                :cvv_message, :error_code, :error_detail, :created_at,
-                :updated_at)
+    attr_reader(
+      :success,
+      :message,
+      :avs_code,
+      :avs_message,
+      :cvv_code,
+      :cvv_message,
+      :error_code,
+      :error_detail,
+      :created_at,
+      :updated_at
+    )
   end
 end